Add our own home-baked big integer implementation.
It is still severely restricted in the features it supports, but implements enough
to compute the Fibonacci numbers. Since  it is a relatively simple implementation,
speed is somewhat lower than more professional options. But it works, and uses no
external crates except num-traits for the definition of the Zero and One traits.
This dependency could be removed as well (the traits are simple enough to easily
reimplement), but they are so well known there seems no upside to doing so.